Implemented my weapon, "Uncommon Bow" and completed answers.txt
Implemented my weapon "uncommon Bow". If the opponenet doesn't have armor the weapon hits for full damage. If the opponenet has armor then the bow has a 25 % chance to deflect and do no damage, 50% chance to graze and do damage based on the armor level, or a 25% chance to pierce the armour and do 80% of the total weapon damage.

Also completed the all of the questions on the answers.txt file.